+++ tests/auto/imageparser/tst_imageparser.cpp
@@ -48,32 +48,32 @@ private Q_SLOTS:
     {
         QTest::addColumn<QString>("filePath");
         QTest::addColumn<bool>("succeeded");
-        QTest::addColumn<int>("imageCount");
+        QTest::addColumn<size_t>("imageCount");
         QTest::addColumn<QString>("imageFile");
 
         QTest::newRow("Valid") << QStringLiteral(ASSETS "imageparser_valid.gltf")
                                << true
-                               << 1
+                               << size_t(1)
                                << ASSETS "aBasePath/anImage.png";
 
         QTest::newRow("Empty") << QStringLiteral(ASSETS "imageparser_empty.gltf")
                                << false
-                               << 0
+                               << size_t(0)
                                << QString();
 
         QTest::newRow("PartiallyIncomplete") << QStringLiteral(ASSETS "imageparser_incomplete.gltf")
                                              << false
-                                             << 0
+                                             << size_t(0)
                                              << QString();
 
         QTest::newRow("UriAndBufferView") << QStringLiteral(ASSETS "imageparser_uri_and_bufferview.gltf")
                                           << false
-                                          << 0
+                                          << size_t(0)
                                           << QString();
 
         QTest::newRow("UriAndBufferView") << QStringLiteral(ASSETS "imageparser_bufferview.gltf")
                                           << false
-                                          << 0
+                                          << size_t(0)
                                           << QString();
     }
 
@@ -81,7 +81,7 @@ private Q_SLOTS:
     {
         QFETCH(QString, filePath);
         QFETCH(bool, succeeded);
-        QFETCH(int, imageCount);
+        QFETCH(size_t, imageCount);
         QFETCH(QString, imageFile);
 
         // GIVEN
@@ -157,7 +157,7 @@ private Q_SLOTS:
 
         // THEN
         QVERIFY(success);
-        QCOMPARE(context.imagesCount(), 1);
+        QCOMPARE(context.imagesCount(), size_t(1));
         Image image = context.image(0);
         QCOMPARE(image.url.scheme(), QStringLiteral("qrc"));
         QCOMPARE(image.url, QUrl("qrc:/anImage.png"));
@@ -180,7 +180,7 @@ private Q_SLOTS:
 
         // THEN
         QVERIFY(success);
-        QCOMPARE(context.imagesCount(), 7);
+        QCOMPARE(context.imagesCount(), size_t(7));
         {
             Image image = context.image(0);
             QCOMPARE(image.url.scheme(), QStringLiteral("file"));
