+++ tests/auto/layerparser/tst_layerparser.cpp
@@ -47,26 +47,26 @@ private Q_SLOTS:
     {
         QTest::addColumn<QString>("filePath");
         QTest::addColumn<bool>("succeeded");
-        QTest::addColumn<int>("layersCount");
+        QTest::addColumn<size_t>("layersCount");
 
         QTest::newRow("Valid") << QStringLiteral(ASSETS "layerparser_valid.gltf")
                                << true
-                               << 3;
+                               << size_t(3);
 
         QTest::newRow("Empty") << QStringLiteral(ASSETS "layerparser_empty.gltf")
                                << false
-                               << 0;
+                               << size_t(0);
 
         QTest::newRow("InvalidContent") << QStringLiteral(ASSETS "layerparser_invalid.gltf")
                                         << false
-                                        << 0;
+                                        << size_t(0);
     }
 
     void checkParse()
     {
         QFETCH(QString, filePath);
         QFETCH(bool, succeeded);
-        QFETCH(int, layersCount);
+        QFETCH(size_t, layersCount);
 
         // GIVEN
         GLTF2Context context;
