+++ tests/auto/materialparser/tst_materialparser.cpp
@@ -54,51 +54,51 @@ private Q_SLOTS:
     {
         QTest::addColumn<QString>("filePath");
         QTest::addColumn<bool>("succeeded");
-        QTest::addColumn<int>("materialCount");
+        QTest::addColumn<size_t>("materialCount");
 
         QTest::newRow("Valid") << QStringLiteral(ASSETS "materialparser_valid.gltf")
                                << true
-                               << 1;
+                               << size_t(1);
 
         QTest::newRow("Empty") << QStringLiteral(ASSETS "materialparser_empty.gltf")
                                << false
-                               << 0;
+                               << size_t(0);
 
         QTest::newRow("Empty Material") << QStringLiteral(ASSETS "materialparser_empty_material.gltf")
                                         << true
-                                        << 1;
+                                        << size_t(1);
 
         QTest::newRow("PartiallyComplete") << QStringLiteral(ASSETS "materialparser_partially_complete.gltf")
                                            << true
-                                           << 1;
+                                           << size_t(1);
 
         QTest::newRow("Invalid_AlphaMode") << QStringLiteral(ASSETS "materialparser_invalid_alpha_mode.gltf")
                                            << false
-                                           << 0;
+                                           << size_t(0);
 
         QTest::newRow("Invalid_BaseColorArray") << QStringLiteral(ASSETS "materialparser_invalid_basecolorarray.gltf")
                                                 << false
-                                                << 0;
+                                                << size_t(0);
 
         QTest::newRow("Invalid_BaseColorArray2") << QStringLiteral(ASSETS "materialparser_invalid_basecolorarray2.gltf")
                                                  << false
-                                                 << 0;
+                                                 << size_t(0);
 
         QTest::newRow("Invalid_EmissiveFactor") << QStringLiteral(ASSETS "materialparser_invalid_emissivefactor.gltf")
                                                 << false
-                                                << 0;
+                                                << size_t(0);
 
         QTest::newRow("Invalid_EmissiveFactor2") << QStringLiteral(ASSETS "materialparser_invalid_emissivefactor2.gltf")
                                                  << false
-                                                 << 0;
+                                                 << size_t(0);
 
         QTest::newRow("KHR_materials_unlit") << QStringLiteral(ASSETS "materialparser_KHR_materials_unlit.gltf")
                                              << true
-                                             << 1;
+                                             << size_t(1);
 #ifdef Q_OS_LINUX
         QTest::newRow("KDAB_custom_material") << QStringLiteral(ASSETS "materialparser_KDAB_custom_material.gltf")
                                               << true
-                                              << 1;
+                                              << size_t(1);
 #endif
     }
 
@@ -112,7 +112,7 @@ private Q_SLOTS:
 
         QFETCH(QString, filePath);
         QFETCH(bool, succeeded);
-        QFETCH(int, materialCount);
+        QFETCH(size_t, materialCount);
 
         // GIVEN
         GLTF2Context context;
@@ -177,7 +177,7 @@ private Q_SLOTS:
 
         // THEN
         QVERIFY(success);
-        QCOMPARE(context.materialsCount(), 1);
+        QCOMPARE(context.materialsCount(), size_t(1));
         QCOMPARE(context.material(0).alpha.alphaCutoff, expectedAlphaCutoff);
         QCOMPARE(int(context.material(0).alpha.mode), expectedBlendMode);
     }
@@ -232,7 +232,7 @@ private Q_SLOTS:
 
         // THEN
         QVERIFY(success);
-        QCOMPARE(context.materialsCount(), 1);
+        QCOMPARE(context.materialsCount(), size_t(1));
         QCOMPARE(context.material(0).pbr.baseColorTexture.khr_texture_transform.offset, offset);
         QCOMPARE(context.material(0).pbr.baseColorTexture.khr_texture_transform.rotation, rotation);
         QCOMPARE(context.material(0).pbr.baseColorTexture.khr_texture_transform.scale, scale);
